Add openTransaction modal
